+++ Sources/StreamChat/Controllers/ChannelController/ChannelController.swift
@@ -150,7 +150,7 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
     /// A boolean value indicating if it should send typing events.
     /// It is `true` if the channel typing events are enabled as well as the user privacy settings.
     func shouldSendTypingEvents(completion: @escaping (Bool) -> Void) {
-        guard channel?.canSendTypingEvents ?? true else {
+        guard channel?.canSendTypingEvents ?? false else {
             completion(false)
             return
         }
@@ -1835,8 +1835,11 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
             return
         }
 
-        /// Send stop typing event.
-        eventSender.stopTyping(in: cid, parentMessageId: nil)
+        /// Send stop typing event only if the channel supports typing events.
+        shouldSendTypingEvents { isEnabled in
+            guard isEnabled else { return }
+            self.eventSender.stopTyping(in: cid, parentMessageId: nil)
+        }
 
         updater.createNewMessage(
             in: cid,